storeの認証に関する情報をstore/authenticationへ移動
