+++ frontend/layouts/default.vue
@@ -13,7 +13,7 @@
           </v-list-item>
         </nuxt-link>
 
-        <nuxt-link to="/user/login" v-if="!$store.state.isAuthenticated">
+        <nuxt-link to="/user/login" v-if="!isAuthenticated">
           <v-list-item link>
             <v-list-item-action>
               <v-icon>mdi-account-arrow-right</v-icon>
@@ -27,7 +27,7 @@
         <v-list-item
           link
           @click="logout"
-          v-if="$store.state.isAuthenticated"
+          v-if="isAuthenticated"
         >
           <v-list-item-action>
             <v-icon>mdi-account-cancel-outline</v-icon>
@@ -37,7 +37,7 @@
           </v-list-item-content>
         </v-list-item>
 
-        <nuxt-link to="/user/sign_up" v-if="!$store.state.isAuthenticated">
+        <nuxt-link to="/user/sign_up" v-if="!isAuthenticated">
           <v-list-item link>
             <v-list-item-action>
               <v-icon>mdi-account-plus</v-icon>
@@ -48,7 +48,7 @@
           </v-list-item>
         </nuxt-link>
 
-        <nuxt-link to="/user/1" v-if="$store.state.isAuthenticated">
+        <nuxt-link to="/user/1" v-if="isAuthenticated">
           <v-list-item link>
             <v-list-item-action>
               <v-icon>mdi-account-details</v-icon>
@@ -70,7 +70,7 @@
           </v-list-item>
         </nuxt-link>
 
-        <nuxt-link to="/memos/write" v-if="$store.state.isAuthenticated">
+        <nuxt-link to="/memos/write" v-if="isAuthenticated">
           <v-list-item link>
             <v-list-item-action>
               <v-icon>mdi-border-color</v-icon>
@@ -117,26 +117,22 @@
 </template>
 
 <script>
-const Cookie = process.client ? require("js-cookie") : undefined;
-
 export default {
-  props: {
-    source: String
-  },
   data: () => ({
     drawer: null
   }),
+
+  computed: {
+    isAuthenticated() {
+      return this.$store.getters["authentication/isAuthenticated"]
+    }
+  },
+
   methods: {
     async logout() {
       try {
-        await this.$store.dispatch("logout", {
-          access_token: Cookie.get("access-token"),
-          client: Cookie.get("client"),
-          uid: Cookie.get("uid")
-        });
-        Cookie.remove("access-token");
-        Cookie.remove("client");
-        Cookie.remove("uid");
+        await this.$store.dispatch("authentication/logout");
+
         this.$router.push(`/user/login`);
       } catch (e) {
         console.error(e);
