+++ frontend/pages/user/edit/_id.vue
@@ -31,8 +31,6 @@
 </template>
 
 <script>
-const Cookie = process.client ? require("js-cookie") : undefined;
-
 export default {
   data: () => ({
     showPassword: false,
