+++ frontend/pages/user/sign_up.vue
@@ -72,6 +72,7 @@ export default {
           email: this.email,
           password: this.password
         });
+
         this.$router.push(`/user/confirm`);
       } catch (error) {
         console.error(error);
