+++ frontend/plugins/axios.js
@@ -1,7 +1,7 @@
 export default function({ app: { $axios }, store }) {
   $axios.onRequest(config => {
-    $axios.setHeader("access-token", store.state.access_token);
-    $axios.setHeader("uid", store.state.uid);
-    $axios.setHeader("client", store.state.client);
+    $axios.setHeader("access-token", store.getters["authentication/access_token"]);
+    $axios.setHeader("uid", store.getters["authentication/uid"]);
+    $axios.setHeader("client", store.getters["authentication/client"]);
   })
 }
